requirements.txt
llama-index==0.10.62
llama-index-vector-stores-qdrant==0.2.14
llama-index-readers-file==0.1.32
llama-index-llms-ollama==0.2.2
llama-index-embeddings-ollama==0.1.3
llama-index-embeddings-fastembed==0.1.7
llama-index-utils-workflow==0.1.0
qdrant-client==1.10.1
python-dotenv==1.0.1
unstructured==0.15.1
fastembed==0.3.1
llama-index-callbacks-arize-phoenix==0.1.6